+++ frontend/src/components/ProxyPromptDialog.vue
@@ -96,15 +96,3 @@ const statusText = computed(() => {
   }
 })
 </script>
-
-<style scoped>
-.dialog-header {
-  background-color: rgb(var(--v-theme-primary));
-  color: white;
-  padding: 16px 24px;
-  font-size: 16px;
-  font-weight: 500;
-  display: flex;
-  align-items: center;
-}
-</style>